+++ doc/source/whats_new.rst
@@ -9,7 +9,7 @@ Current
 Changelog
 ~~~~~~~~~
 
-   - ...
+   - Linear (and zeroth-order) detrending for Epochs and Evoked by `Eric Larson`_
 
 
 .. _changes_0_5: